requirements.txt
asgiref==3.8.1
bleach==6.1.0
customtkinter==5.2.2
darkdetect==0.8.0
Django==4.2.11
django-extensions==3.2.3
django-highlights==0.0.8
django-summernote==0.8.20.0
django_debug_toolbar==3.8.1
gunicorn==22.0.0
Markdown
packaging==24.0
pillow 
psycopg2==2.9.9
pydub==0.25.1
Pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
webencodings==0.5.1
whitenoise==6.6.0













